# api/openapi.yaml
openapi: 3.0.3
info:
  title: Auth Service API
  description: Authentication service providing user login, token refresh, logout, and user information endpoints
  version: 1.0.0
  contact:
    name: Auth Service Team
    email: support@authservice.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8090
    description: Local development server
  - url: https://e29d425094dc.ngrok-free.app/
    description: Production server
paths:
...
  /v1/auth/refresh:
    post:
      summary: Refresh access token
      description: Generate new access token using refresh token
      operationId: refresh
      tags:
        - Authentication
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      parameters:
        - name: Refresh-Token
          in: header
          description: Refresh token from header (alternative to cookie)
          required: false
          schema:
            type: string
            example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        - name: Cookie
          in: header
          description: Refresh token from cookie
          required: false
          schema:
            type: string
            example: refresh_token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Token refresh successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token for authentication
  schemas:
...
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          description: Refresh token (optional if provided via cookie)
    RefreshResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            access_token:
              type: string
              example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
              description: New JWT access token
            token_type:
              type: string
              example: "Bearer"
              description: Token type
            expires_in:
              type: integer
              example: 3600
              description: Token expiration time in seconds
          required:
            - access_token
            - token_type
            - expires_in
        success:
          type: boolean
          example: true
      required:
        - data
        - success
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "INVALID_CREDENTIALS"
              description: Error code
            message:
              type: string
              example: "Invalid email or password"
              description: Human-readable error message
            details:
              type: object
              nullable: true
              description: Additional error details
          required:
            - code
            - message
        success:
          type: boolean
          example: false
      required:
        - error
        - success
...